# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TuckZone Canteen API",
        "description": "Backend for the school canteen pre-order app.\n\nAuthentication: obtain tokens from /api/auth/login or\n/api/auth/otp/login, then send `Authorization: Bearer <accessToken>`.\nWhen a call returns 401, exchange the refresh token at\n/api/auth/refresh rather than forcing the user to sign in again.\n\nOrdering is idempotent: generate one idempotencyKey when the\ncheckout screen opens and reuse it for every retry of that\nsame order, otherwise a double tap creates two orders.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/wallet/topup": {
            "post": {
                "tags": [
                    "wallet-controller"
                ],
                "operationId": "initiateTopup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TopupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TopupInitResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/wallet/topup/verify": {
            "post": {
                "tags": [
                    "wallet-controller"
                ],
                "operationId": "verifyTopup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyTopupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/wallet/topup/mock-complete": {
            "post": {
                "tags": [
                    "wallet-controller"
                ],
                "operationId": "mockCompleteTopup",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MockTopupCompleteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/users/{id}/enable": {
            "post": {
                "tags": [
                    "user-admin-controller"
                ],
                "operationId": "enable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/users/{id}/disable": {
            "post": {
                "tags": [
                    "user-admin-controller"
                ],
                "operationId": "disable",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/expenses": {
            "get": {
                "tags": [
                    "report-controller"
                ],
                "operationId": "listExpenses",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExpenseResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "report-controller"
                ],
                "operationId": "addExpense",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpenseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpenseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/wallet/transactions": {
            "get": {
                "tags": [
                    "wallet-controller"
                ],
                "operationId": "transactions",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WalletTransactionResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/users": {
            "get": {
                "tags": [
                    "user-admin-controller"
                ],
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "STUDENT",
                                "PARENT",
                                "TEACHER",
                                "CANTEEN_ADMIN"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserSummary"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/reports/sales": {
            "get": {
                "tags": [
                    "report-controller"
                ],
                "operationId": "sales",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalesReportResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/dashboard": {
            "get": {
                "tags": [
                    "report-controller"
                ],
                "operationId": "dashboard",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DashboardResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/expenses/{id}": {
            "delete": {
                "tags": [
                    "report-controller"
                ],
                "operationId": "deleteExpense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserSummary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "STUDENT",
                            "PARENT",
                            "TEACHER",
                            "CANTEEN_ADMIN"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ACTIVE",
                            "DISABLED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "admissionNumber": {
                        "type": "string"
                    },
                    "studentClass": {
                        "type": "string"
                    },
                    "section": {
                        "type": "string"
                    },
                    "rollNumber": {
                        "type": "string"
                    },
                    "seatNumber": {
                        "type": "string"
                    },
                    "studentMobile": {
                        "type": "string"
                    },
                    "parentMobile": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    }
                }
            },
...
            "TopupRequest": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number",
                        "minimum": 1.0
                    }
                },
                "required": [
                    "amount"
                ]
            },
            "TopupInitResponse": {
                "type": "object",
                "properties": {
                    "topupId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "gatewayOrderId": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "gatewayKeyId": {
                        "type": "string"
                    }
                }
            },
            "VerifyTopupRequest": {
                "type": "object",
                "properties": {
                    "gatewayOrderId": {
                        "type": "string",
                        "minLength": 1
                    },
                    "gatewayPaymentId": {
                        "type": "string",
                        "minLength": 1
                    },
                    "signature": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "gatewayOrderId",
                    "gatewayPaymentId",
                    "signature"
                ]
            },
            "WalletResponse": {
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "balance": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    }
                }
            },
            "MockTopupCompleteRequest": {
                "type": "object",
                "properties": {
                    "gatewayOrderId": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "gatewayOrderId"
                ]
            },
...
            "ExpenseRequest": {
                "type": "object",
                "properties": {
                    "expenseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "INGREDIENTS",
                            "STAFF_WAGES",
                            "GAS_FUEL",
                            "RENT",
                            "UTILITIES",
                            "PACKAGING",
                            "EQUIPMENT",
                            "OTHER"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "amount": {
                        "type": "number",
                        "minimum": 0.01
                    }
                },
                "required": [
                    "amount",
                    "category",
                    "expenseDate"
                ]
            },
            "ExpenseResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "expenseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "INGREDIENTS",
                            "STAFF_WAGES",
                            "GAS_FUEL",
                            "RENT",
                            "UTILITIES",
                            "PACKAGING",
                            "EQUIPMENT",
                            "OTHER"
                        ]
                    },
                    "description": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    }
                }
            },
...
            "WalletTransactionResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "CREDIT",
                            "DEBIT"
                        ]
                    },
                    "amount": {
                        "type": "number"
                    },
                    "balanceAfter": {
                        "type": "number"
                    },
                    "referenceType": {
                        "type": "string"
                    },
                    "referenceId": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "DailySalesRow": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "orders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "revenue": {
                        "type": "number"
                    }
                }
            },
...
            "SalesReportResponse": {
                "type": "object",
                "properties": {
                    "from": {
                        "type": "string",
                        "format": "date"
                    },
                    "to": {
                        "type": "string",
                        "format": "date"
                    },
                    "revenue": {
                        "type": "number"
                    },
                    "costOfGoods": {
                        "type": "number"
                    },
                    "grossProfit": {
                        "type": "number"
                    },
                    "expenses": {
                        "type": "number"
                    },
                    "netProfit": {
                        "type": "number"
                    },
                    "orderCount": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "daily": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DailySalesRow"
                        }
                    },
                    "topItems": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TopItemRow"
                        }
                    },
                    "peakHours": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PeakHourRow"
                        }
                    }
                }
            },
            "TopItemRow": {
                "type": "object",
                "properties": {
                    "itemName": {
                        "type": "string"
                    },
                    "quantitySold": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "revenue": {
                        "type": "number"
                    }
                }
            },
...
            "DashboardResponse": {
                "type": "object",
                "properties": {
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "totalOrders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pendingOrders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "completedOrders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rejectedOrders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "cancelledOrders": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "revenue": {
                        "type": "number"
                    },
                    "costOfGoods": {
                        "type": "number"
                    },
                    "grossProfit": {
                        "type": "number"
                    },
                    "expenses": {
                        "type": "number"
                    },
                    "netProfit": {
                        "type": "number"
                    },
                    "totalCustomers": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "topItems": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TopItemRow"
                        }
                    },
                    "lowStock": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LowStockRow"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}